requirements.txt
click==7.1.2
cycler==0.10.0
dnspython==1.16.0
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.15.1
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
numpy==1.18.4
pandas==1.0.3
pymongo==3.10.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
scikit-learn==0.23.0
scipy==1.4.1
seaborn==0.10.1
six==1.14.0
sklearn==0.0
threadpoolctl==2.0.0
Werkzeug==1.0.1
